+++ py/desisurvey/schedule.py
@@ -665,7 +665,7 @@ def initialize(ephem, start_date=None, stop_date=None, step_size=5.0,
     calendar['monsoon'] = np.zeros(num_nights, bool)
     calendar['fullmoon'] = np.zeros(num_nights, bool)
     calendar['weather'] = np.zeros(num_nights, np.float32)
-    weather_weights = 1 - desisurvey.utils.dome_closed_probabilities()
+    weather_weights = 1 - desisurvey.utils.dome_closed_fractions()
 
     # Prepare a table of ephemeris data.
     etable = astropy.table.Table()
@@ -712,7 +712,7 @@ def initialize(ephem, start_date=None, stop_date=None, step_size=5.0,
         calendar[i]['monsoon'] = desisurvey.utils.is_monsoon(midnight[i])
         calendar[i]['fullmoon'] = ephem.is_full_moon(midnight[i])
         # Look up expected dome-open fraction due to weather.
-        calendar[i]['weather'] = weather_weights[date.month - 1]
+        calendar[i]['weather'] = weather_weights[i]
         # Calculate the program during this night (default is 4=DAYTIME).
         mjd = midnight[i] + t_centers
         dark, gray, bright = ephem.get_program(mjd)